renamed import to match actual create_handler
