+++ bin/lfg_bot.dart
@@ -9,8 +9,8 @@ import 'package:lfg_bot/core/l10n/generated/messages_all_locales.dart';
 import 'package:lfg_bot/core/utils/context/context.dart';
 import 'package:lfg_bot/core/utils/database/tables/posts.dart';
 import 'package:lfg_bot/core/utils/loaders/bot_settings.dart';
-import 'package:lfg_bot/features/admin/test.dart';
-import 'package:lfg_bot/features/create/handler/create_handle.dart';
+import 'package:lfg_bot/features/admin/admin_commands_creator.dart';
+import 'package:lfg_bot/features/create/handler/create_handler.dart';
 import 'package:lfg_bot/features/delete/handler/delete_handler.dart';
 import 'package:lfg_bot/features/edit/handler/edit_handler.dart';
 import 'package:lfg_bot/features/join/handler/join_handle.dart';
